var type changed on ft_negint

+++ libformat.c
@@ -72,11 +72,11 @@ int		ft_padding(t_prtform *form)
 
 char	*ft_negint(t_prtform *form, int i, char *s)
 {
-	char	*temp;
-	int		c;
+	char				*temp;
+	unsigned long int	c;
 
 	c = 0;
-	while (i < 0 && form->pad_char == '0' && (c + ft_strlen(s) + 1) < (int)form->padding)
+	while (i < 0 && form->pad_char == '0' && (c + ft_strlen(s) + 1) < form->padding)
 	{
 		temp = ft_strjoin("0", s);
 		free(s);
